This is the first version of DAT 510 Milestone 2
This is the first version of my decision tree.  I also have loaded the 2 csv files from class, created another csv which is the decision tree predictions.  Also, there is an Rmd file to visualize the frist version of the script.  Initial results are good, though there may be some over fitting as some of the resulte indicate 100% yes or 100% no.  I am yet to determine the cause for this.